--- infrastructure/test-docker-compose.yml
@@ -1,34 +1,0 @@
-version: "3"
-
-services:
-  postgres_db:
-    image: postgres:14.5
-    restart: always
-    environment:
-      - POSTGRES_USER=postgres
-      - POSTGRES_PASSWORD=postgres
-    ports:
-      - "5432:5432"
-
-  tests:
-    build:
-      context: ../
-      dockerfile: infrastructure/Dockerfile
-      args:
-        TESTING: 1
-    environment:
-      - DB_URL_PV=postgresql://postgres:postgres@postgres_db:5432/postgres
-      - GIT_PYTHON_REFRESH=quiet
-      - LOG_LEVEL=DEBUG
-    command: >
-      bash -c "source $$(poetry env info --path)/bin/activate
-      && pytest --cov=./pv_site_production
-      && coverage report -m
-      && coverage xml
-      && cp .coverage ./tests/
-      && cp coverage.xml ./tests/"
-    depends_on:
-      - postgres_db
-    volumes:
-      - ../pv_site_production/:/app/pv_site_production
-      - ../tests/:/app/tests
