+++ pv_site_production/models/psp/psp_model.py
@@ -1,13 +1,13 @@
 from typing import Any
 
 import pandas as pd
-from pv_site_production.utils.imports import instantiate
-
 from psp.data.data_sources.pv import PvDataSource
 
 # TODO This should live in psp.ml.models.base
 from psp.ml.models.recent_history import SetupConfig
 from psp.ml.serialization import load_model
+
+from pv_site_production.utils.imports import instantiate
 
 
 def get_model(config: dict[str, Any], pv_data_source: PvDataSource) -> pd.DataFrame:
