--- tests/models/nwp_pv/test_nwp_pv_load.py
@@ -1,8 +1,0 @@
-from pv_site_production.models.pv_nwp.load import load_model
-
-
-def test_load_model():
-    model = load_model()
-    # TODO
-    print(model)
-    # assert model is not None
